# Support OpenGL < 4.2 in PBS shader