--- a/ogre2/src/media/Hlms/Pbs/GLSL/BRDFs_piece_ps.glsl
+++ b/ogre2/src/media/Hlms/Pbs/GLSL/BRDFs_piece_ps.glsl
@@ -12,7 +12,7 @@
 @property( !fresnel_scalar )
 	@piece( maxR1F0 )max( 1.0 - ROUGHNESS, @insertpiece( F0 ).x )@end
 @end @property( fresnel_scalar )
-	@piece( maxR1F0 )max( (1.0 - ROUGHNESS).xxx, @insertpiece( F0 ).xyz )@end
+	@piece( maxR1F0 )max( (1.0 - vec3(ROUGHNESS, 0.0, 0.0)).xxx, @insertpiece( F0 ).xyz )@end
 @end
 
 //For mortals: